Crate muoxi_watchdog

WatchDog that monitors the custom defined .json files located within config/ directory Runs as a completely seperate process apart from all servers. Watchdogs main job is to notice contents changes of file and sync them with Database

Basic usage to watch a file

let mut watchdog = Hotwatch::new_with_custom_delay(Duration::from_millis(100)?);

watcher.watch("config/accounts.json", move |event|{
    if let Event::Write(_path) = event{
        // sync accounts.json with accounts database in MongoDB
        trigger_upload(JsonFile::Accounts).unwrap();
    }
})?;

Enums

JsonFile
Different .json storage files that need to be monitored

Statics

CLIENTS
static path to place of clients json file

Functions

main

read_file
simple wrapper to read from json file and return serde_json::Value

trigger_upload
main function that triggers upload protocols for each change in file based on JsonFile